+++ src/screens/InputTelNumberScreen/InputTelNumberScreen.tsx
@@ -41,14 +41,14 @@ const InputTelNumberScreen: React.FC<InputOtpScreenNavigationProp> = ({ navigati
       alert('กรุณาใส่หมายเลขโทรศัพท์')
     } else {
       const telephoneNo = fillZero(tel)
-      VerifiesDataSource.verifyPhoneNo(telephoneNo).then((res:ResponseEntity<OtpRequestEntity>) => {
+      VerifiesDataSource.verifyPhoneNo(telephoneNo).then((res: ResponseEntity<OtpRequestEntity>) => {
         if (res.success) {
           setIsError(false)
-          // if (bypassTelephone.includes(tel)) {
-          //   navigation.navigate('LoginSuccess', { userProfile: res })
-          // } else {
-          navigation.navigate('InputOtp', { data: res.responseData })
-          // }
+          if (bypassTelephone.includes(tel)) {
+            navigation.navigate('LoginSuccess', { userProfile: res })
+          } else {
+            navigation.navigate('InputOtp', { data: res.responseData })
+          }
         } else {
           setIsError(true)
           setMessage(res.userMessage)